fix: filename_removal.fish -> ファイル名が空の場合と既に存在している場合上書きを防ぐために処理をスキップする.

+++ default/.config/fish/user/filename_removal.fish
@@ -10,6 +10,19 @@ function filename_removal
         if test -f "$file"
             set new_name (string replace -a "$target" "" -- "$file")
 
+            # ファイル名が空の場合スキップ.
+            if test -z "$new_name"
+                echo "Skip '$file' -> '$new_name' due to empty filename."
+                continue
+            end
+
+            # ファイル名が既に存在しているため.
+            # ファイルの上書きを防ぐ.
+            if test -e "$new_name"
+                echo "Skip '$file' => '$new_name' due to existing already."
+                continue
+            end
+
             if test "$file" != "$new_name"
                 echo "$file => $new_name"
                 mv -- "$file" "$new_name"
